Edited and proofread.
Hello! I added and deleted words to fix awkward sentences. I changed uppercase letters to lowercase because the words after comma doesn't need to be capitalized unless it's a proper noun. I deleted the question mark in How Nix Works. I have one suggestion for the title, it would be better as Getting Started With Nix: Introduction... I didn't change it here because I don't want to screw up to code.

Overall, it's a fitting post for an introduction. Short and sweet (: there aren't really much to change which is great. I love that you used a diagram too. Great work, keep it up! In the next posts of the series, it would be great to also incorporate some GIFs if needed.